:sparkles: [Feat] 받은 요청 페이지 기본 틀 구현
