+++ src/page/driver/costCall/components/CallList.tsx
@@ -1,0 +1,5 @@
+import style from './CallList.module.css';
+
+export default function CallList() {
+  return <div className={style.callList}>CallList</div>;
+}
